refactor(servlet): overall cleansing packages
