+++ tomcat/src/main/java/nextstep/jwp/controller/GetController.java
@@ -1,6 +1,6 @@
 package nextstep.jwp.controller;
 
-import nextstep.jwp.servlet.RequestMapping;
+import nextstep.jwp.servlet.handler.RequestMapping;
 import org.apache.coyote.support.HttpMethod;
 
 public class GetController {
