+++ tomcat/src/main/java/nextstep/jwp/servlet/handler/RequestMapping.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.servlet;
+package nextstep.jwp.servlet.handler;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
