+++ tomcat/src/main/java/nextstep/jwp/servlet/view/ViewResolver.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.servlet;
+package nextstep.jwp.servlet.view;
 
 import java.io.File;
 import java.io.IOException;
